fixes & optimizations

+++ hybrid.sh
@@ -25,9 +25,9 @@ nc='\033[0m'
 
 body(){
 	clear
-	echo "${cyan}[-=The Hybrid Project=-]${nc}"
+	echo "$cyan[-=The Hybrid Project=-]$nc"
 	echo
-	echo "${yellow}Menu:${nc}"
+	echo "${yellow}Menu:$nc"
 	echo " 1|Instant Boost"
 	echo " 2|Clean up my crap"
 	echo " 3|Optimize my SQLite DB's"
@@ -35,7 +35,10 @@ body(){
 	echo " 5|Tune my LMK"
 	echo " 6|Tune my Networks"
 	echo " 7|Kernel Kontrol"
-	echo " 8|zRAM Settings"
+	if [ -f /dev/block/zram* ]
+	then
+	 echo " 8|zRAM Settings"
+	fi
 	echo " 9|Game Booster"
 	echo
 	echo " O|Options"
@@ -70,43 +73,50 @@ unknown_option(){
 }
 
 init_sleep(){
-	if [ ! -f system/etc/init.d/50sleep ]; then
+	if [ ! -f system/etc/init.d/50sleep ]
+	then
 		touch /system/etc/init.d/50sleep
 		chmod 755 /system/etc/init.d/50sleep
 		echo -ne "" > /system/etc/init.d/50sleep
 cat >> /system/etc/init.d/50sleep <<EOF
 #!/system/bin/sh
+
 sleep 10
+
 EOF
 	fi
 }
 
 drop_caches(){
 	clear
-	echo "${yellow}Dropping caches...${nc}"
+	echo "${yellow}Dropping caches...$nc"
 	sleep 1
 	sync
 	echo "3" > /proc/sys/vm/drop_caches
 	clear
-	echo "${yellow}Caches dropped!${nc}"
+	echo "${yellow}Caches dropped!$nc"
 	sleep 1
-	if [ $perm == 1 ] && [ $initd == 1 ]; then
+	if [ $perm == 1 ] && [ $initd == 1 ]
+	then
 		init_sleep
 		touch /system/etc/init.d/97cache_drop
 		chmod 755 /system/etc/init.d/97cache_drop
 		echo -ne "" > /system/etc/init.d/97cache_drop
 cat >> /system/etc/init.d/97cache_drop <<EOF
 #!/system/bin/sh
-sync && echo "3" > /proc/sys/vm/drop_caches
+
+sync
+echo "3" > /proc/sys/vm/drop_caches
+
 EOF
-	echo "${yellow}Installed!${nc}"
+	echo "${yellow}Installed!$nc"
 	sleep 1
 	fi
 }
 
 clean_up(){
 	clear
-	echo "${yellow}Cleaning up...${nc}"
+	echo "${yellow}Cleaning up...$nc"
 	sleep 1
 
 	#cleaner
@@ -130,12 +140,13 @@ clean_up(){
 	rm -f /data/tombstones/*
 	rm -f /data/system/dropbox/*
 	rm -f /data/system/usagestats/*
-	rm -rf $EXTERNAL_STORAGE/LOST.DIR/
+	rm -f $EXTERNAL_STORAGE/LOST.DIR/*
 
 	clear
-	echo "${yellow}Clean up complete!${nc}"
+	echo "${yellow}Clean up complete!$nc"
 	sleep 1
-	if [ $perm == 1 ] && [ $initd == 1 ]; then
+	if [ $perm == 1 ] && [ $initd == 1 ]
+	then
 		init_sleep
 		touch /system/etc/init.d/99clean_up
 		chmod 755 /system/etc/init.d/99clean_up
@@ -163,52 +174,57 @@ rm -f /data/mlog/*
 rm -f /data/tombstones/*
 rm -f /data/system/dropbox/*
 rm -f /data/system/usagestats/*
-rm -rf $EXTERNAL_STORAGE/LOST.DIR/
+rm -f $EXTERNAL_STORAGE/LOST.DIR/*
 
 EOF
-	echo "${yellow}Installed!${nc}"
+	echo "${yellow}Installed!$nc"
 	sleep 1
 	fi
 }
 
 sql_optimize(){
 	clear
-	echo "${yellow}Optimizing SQLite databases...${nc}"
-	echo
+	echo "${yellow}Optimizing SQLite databases...$nc"
 	sleep 1
+	clear
 
-	if [ -e /system/xbin/sqlite3 ]; then
+	if [ -e /system/xbin/sqlite3 ]
+	then
 		chown root.root  /system/xbin/sqlite3
 		chmod 755 /system/xbin/sqlite3
 		SQLLOC=/system/xbin/sqlite3
 	fi
 
-	if [ -e /system/bin/sqlite3 ]; then
+	if [ -e /system/bin/sqlite3 ]
+	then
 		chown root.root /system/bin/sqlite3
 		chmod 755 /system/bin/sqlite3
 		SQLLOC=/system/bin/sqlite3
 	fi
 
-	if [ -e /system/sbin/sqlite3 ]; then
+	if [ -e /system/sbin/sqlite3 ]
+	then
 		chown root.root /sbin/sqlite3
 		chmod 755 /sbin/sqlite3
 		SQLLOC=/sbin/sqlite3
 	fi
-	for i in `find ./ -iname "*.db"`; do
-		$SQLLOC $i 'VACUUM;'
-		clear; echo "${yellow}Vacuumed:${nc} $i"
-		$SQLLOC $i 'REINDEX;'
-		echo "${yellow}Reindexed :${nc} $i"
+	for i in `find / -iname "*.db" 2>/dev/null`
+	do
+		$SQLLOC $i 'VACUUM'
+		echo "${yellow}Vacuumed:$nc $i"
+	 	clear
+		$SQLLOC $i 'REINDEX'
+		echo "${yellow}Reindexed:$nc $i"
 	done
 
 	clear
-	echo "${yellow}SQLite database optimizations complete!${nc}"
+	echo "${yellow}SQLite database optimizations complete!$nc"
 	sleep 1
 }
 
 vm_tune(){
 	clear
-	echo "${yellow}Optimizing VM...${nc}"
+	echo "${yellow}Optimizing VM...$nc"
 	sleep 1
 		
 	echo "80" > /proc/sys/vm/swappiness
@@ -223,9 +239,10 @@ vm_tune(){
 	echo "1" > /proc/sys/vm/oom_kill_allocating_task
 		
 	clear
-	echo "${yellow}VM Optimized!${nc}"
+	echo "${yellow}VM Optimized!$nc"
 	sleep 1
-	if [ $perm == 1 ] && [ $initd == 1 ]; then
+	if [ $perm == 1 ] && [ $initd == 1 ]
+	then
 		init_sleep
 		touch /system/etc/init.d/75vm
 		chmod 755 /system/etc/init.d/75vm
@@ -244,51 +261,55 @@ echo "150" > /proc/sys/vm/overcommit_ratio
 echo "4096" > /proc/sys/vm/min_free_kbytes
 echo "1" > /proc/sys/vm/oom_kill_allocating_task
 EOF
-		echo "${yellow}Installed!${nc}"
+		echo "${yellow}Installed!$nc"
 		sleep 1
 	fi
 }
 
 lmk_tune_opt(){
 	clear
-	echo "${yellow}LMK Optimization!${nc}"
+	echo "${yellow}LMK Optimization$nc"
 	echo
-	echo "${yellow}Minfree profiles available:${nc}"
+	echo "${yellow}Minfree profiles available:$nc"
 	echo " B|Balanced"
 	echo " M|Multitasking|"
 	echo " G|Gaming"
 	echo -n "> "
 	read lmk_opt
 	case $lmk_opt in
-		b|B|m|M|g|G ) echo "Ok!" && sleep 1 && lmk_profile=$lmk_opt && lmk_apply;;
-		* ) error_404 && lmk_tune_opt;;
+		b|B|m|M|g|G ) clear; lmk_profile=$lmk_opt; lmk_apply; echo "Done"; sleep 1;;
+		* ) unknown_option; lmk_tune_opt;;
 	esac
 }
 
 lmk_apply(){
 	clear
-	if [ $lmk_profile == b ] || [ $lmk_profile = B ]; then
+	if [ $lmk_profile == b ] || [ $lmk_profile = B ]
+	then
 		minfree_array='1024,2048,4096,8192,12288,16384'
 	fi
 
-	if [ $lmk_profile == m ] || [ $lmk_profile = M ]; then
-		minfree_array='1536,2048,4096,5120,5632,6144'
+	if [ $lmk_profile == m ] || [ $lmk_profile = M ]
+	then
+	 minfree_array='1536,2048,4096,5120,5632,6144'
 	fi
 
-	if [ $lmk_profile == g ] || [ $lmk_profile = G ]; then
-		minfree_array='10393,14105,18188,27468,31552,37120'
+	if [ $lmk_profile == g ] || [ $lmk_profile = G ]
+	then
+	 minfree_array='10393,14105,18188,27468,31552,37120'
 	fi
 
-	echo "${yellow}Optimizing LMK...${nc}"
+	echo "${yellow}Optimizing LMK...$nc"
 	sleep 1
 		
 	echo "$minfree_array" > /sys/module/lowmemorykiller/parameters/minfree
 		
 	clear
-	echo "${yellow}LMK Optimized!${nc}"
+	echo "${yellow}LMK Optimized!$nc"
 	sleep 1
 
-	if [ $perm == 1 ] && [ $initd == 1 ]; then
+	if [ $perm == 1 ] && [ $initd == 1 ]
+	then
 		init_sleep
 		touch /system/etc/init.d/95lmk
 		chmod 755 /system/etc/init.d/95lmk
@@ -298,14 +319,14 @@ cat >> /system/etc/init.d/95lmk <<EOF
 
 echo "$minfree_array" > /sys/module/lowmemorykiller/parameters/minfree
 EOF
-		echo "${yellow}Installed!${nc}"
+		echo "${yellow}Installed!$nc"
 		sleep 1
 	fi
 }
 
 network_tune(){
 	clear
-	echo "${yellow}Optimizing Networks...${nc}"
+	echo "${yellow}Optimizing Networks...$nc"
 	sleep 1
 
 	#TCP
@@ -350,10 +371,11 @@ network_tune(){
 	echo "1" > /proc/sys/net/ipv4/conf/default/log_martians
 
 	clear
-	echo "${yellow}Networks Optimized!${nc}"
+	echo "${yellow}Networks Optimized!$nc"
 	sleep 1
 
-	if [ $perm == 1 ] && [ $initd == 1 ]; then
+	if [ $perm == 1 ] && [ $initd == 1 ]
+	then
 		init_sleep
 		touch /system/etc/init.d/56net
 		chmod 755 /system/etc/init.d/56net
@@ -401,19 +423,21 @@ echo "0" > /proc/sys/net/ipv4/conf/default/accept_source_route
 echo "1" > /proc/sys/net/ipv4/conf/all/log_martians
 echo "1" > /proc/sys/net/ipv4/conf/default/log_martians
 EOF
-		echo "${yellow}Installed!${nc}"
+		echo "${yellow}Installed!$nc"
 		sleep 1
 	fi
 }
 
 kernel_kontrol(){
 	clear
-	echo "${yellow}Kernel Kontrol${nc}"
-	echo
+	echo "${yellow}Kernel Kontrol$nc"
 	echo " 1|Set CPU Freq"
 	echo " 2|Set CPU Gov"
 	echo " 3|Set I/O Sched"
-	echo " 4|View KCal Values"
+	if [ -d /sys/devices/platform/kcal_ctrl.0/ ]
+	then
+	 	echo " 4|View KCal Values"
+	fi
 	echo " B|Back"
 	echo -n "> "
 	read kk_opt
@@ -423,7 +447,7 @@ kernel_kontrol(){
 		3) setiosched;;
 		4) kcal;;
 		b|B) body;;
-		* ) unknown_option && kernel_kontrol;;
+		* ) unknown_option; kernel_kontrol;;
 	 esac
 }
 
@@ -435,23 +459,23 @@ setcpufreq(){
 	curfreq=`cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_cur_freq`
 	listfreq=`cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_available_frequencies`
 
-	echo "${yellow}CPU Control${nc}"
+	echo "${yellow}CPU Control$nc"
 	echo
-	echo "${bld}Max Freq:${nc} $maxfreq"
-	echo "${bld}Min Freq:${nc} $minfreq"
-	echo "${bld}Current Freq:${nc} $curfreq"
+	echo "${bld}Max Freq:$nc $maxfreq"
+	echo "${bld}Min Freq:$nc $minfreq"
+	echo "${bld}Current Freq:$nc $curfreq"
 	echo
-	echo "${bld}Available Freq's:${nc} "
+	echo "${bld}Available Freq's:$nc"
 	echo "$listfreq"
 	echo
-	echo -n "New Max Freq: " && read newmaxfreq
-	echo -n "New Min Freq: " && read newminfreq
+	echo -n "New Max Freq: "; read newmaxfreq
+	echo -n "New Min Freq: "; read newminfreq
 
 	echo "$newmaxfreq" > /sys/devices/system/cpu/cpu0/cpufreq/scaling_max_freq
 	echo "$newminfreq" > /sys/devices/system/cpu/cpu0/cpufreq/scaling_min_freq
 
 	clear
-	echo "${yellow}New Freq's applied!${nc}"
+	echo "${yellow}New Freq's applied!$nc"
 	sleep 1
 
 	kernel_kontrol
@@ -464,19 +488,19 @@ setgov(){
 	curgov=`cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_governor`
 	listgov=`cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_available_governors`
 
-	echo "${yellow}Governor Control${nc}"
+	echo "${yellow}Governor Control$nc"
 	echo
-	echo "${bld}Current Governor:${nc} $curgov"
+	echo "${bld}Current Governor:$nc $curgov"
 	echo
-	echo "${bld}Available Governors:${nc} "
+	echo "${bld}Available Governors:$nc"
 	echo "$listgov"
 	echo
-	echo -n "New Governor: " && read newgov
+	echo -n "New Governor: "; read newgov
 
 	echo "$newgov" > /sys/devices/system/cpu/cpu0/cpufreq/scaling_governor
 
 	clear
-	echo "${yellow}New Governor applied!${nc}"
+	echo "${yellow}New Governor applied!$nc"
 	sleep 1
 
 	kernel_kontrol
@@ -490,21 +514,22 @@ setiosched(){
 	curiosched=`cat /sys/block/mmcblk0/queue/scheduler | sed 's/.*\[\([a-zA-Z0-9_]*\)\].*/\1/'`
 	listiosched=`cat /sys/block/mmcblk0/queue/scheduler | tr -s "[[:blank:]]" "\n" | sed 's/\[\([a-zA-Z0-9_]*\)\]/\1/'`
 
-	echo "${yellow}I/O Schedulder Control${nc}"
+	echo "${yellow}I/O Schedulder Control$nc"
 	echo
-	echo "${bld}Current I/O Scheduler:${nc} $curiosched"
+	echo "${bld}Current I/O Scheduler:$nc $curiosched"
 	echo
-	echo "${bld}Available I/O Schedulers:${nc} "
+	echo "${bld}Available I/O Schedulers:$nc"
 	echo "$listiosched"
 	echo
-	echo -n "New Scheduler: " && read newiosched
+	echo -n "New Scheduler: "; read newiosched
 
-	for j in /sys/block/*/queue/scheduler; do
+	for j in /sys/block/*/queue/scheduler
+	do
 		echo "$newiosched" > $j
 	done
 
 	clear
-	echo "${yellow}New I/O Scheduler applied!${nc}"
+	echo "${yellow}New I/O Scheduler applied!$nc"
 	sleep 1
 
 	kernel_kontrol
@@ -512,50 +537,59 @@ setiosched(){
 
 kcal(){
 	clear
-	if [ -d /sys/devices/platform/kcal_ctrl.0 ]; then
-		echo "${yellow}Current KCal Values:${nc}"
-		rgb=`cat /sys/devices/platform/kcal_ctrl.0/kcal`
-		sat=`cat /sys/devices/platform/kcal_ctrl.0/kcal_sat`
-		cont=`cat /sys/devices/platform/kcal_ctrl.0/kcal_cont`
-		hue=`cat /sys/devices/platform/kcal_ctrl.0/kcal_hue`
-		gamma=`cat /sys/devices/platform/kcal_ctrl.0/kcal_val`
-		echo "rgb: $rgb, sat: $sat, cont: $cont, hue: $hue, gamma: $gamma"
+	if [ ! -d /sys/devices/platform/kcal_ctrl.0/ ]
+	then
+	 	unknown_option
+	 	kernel_kontrol
 	else
-		echo "KCal driver is missing"
-	fi
-	sleep 5
+	 	echo "${yellow}Current KCal Values:${nc}"
+	 	rgb=`cat /sys/devices/platform/kcal_ctrl.0/kcal`
+	 	sat=`cat /sys/devices/platform/kcal_ctrl.0/kcal_sat`
+	 	cont=`cat /sys/devices/platform/kcal_ctrl.0/kcal_cont`
+	 	hue=`cat /sys/devices/platform/kcal_ctrl.0/kcal_hue`
+	 	gamma=`cat /sys/devices/platform/kcal_ctrl.0/kcal_val`
+	 	echo "rgb: $rgb, sat: $sat, cont: $cont, hue: $hue, gamma: $gamma"
+	 	sleep 5
 
-	kernel_kontrol
+	 	kernel_kontrol
+	fi
 }
 
 zram_settings(){
 	clear
-	echo "${yellow}zRAM Options:${nc}"
-	echo " 1|Disable zRAM"
-	echo " 2|Enable zRAM"
-	echo " B|Back"
-	echo
-	echo -n "> "
-	read options_opt
-	case $options_opt in
-		1 ) zram_disable;;
-		2 ) zram_enable;;
-		b|B ) body;;
-		* ) unknown_option && zram_settings;;
-	esac
+	if [ ! -f /dev/block/zram* ]
+	then
+	 	unknown_option
+	else
+	 	echo "${yellow}zRAM Options:$nc"
+	 	echo
+	 	echo " 1|Disable zRAM"
+	 	echo " 2|Enable zRAM"
+	 	echo " B|Back"
+	 	echo
+	 	echo -n "> "
+	 	read options_opt
+	 	case $options_opt in
+	 		1 ) zram_disable;;
+	 		2 ) zram_enable;;
+	 		b|B ) body;;
+	 		* ) unknown_option; zram_settings;;
+	 	esac
+	fi
 }
 
 zram_enable(){
 	clear
-	echo "${yellow}Enabling zRAM...${nc}"
+	echo "${yellow}Enabling zRAM...$nc"
 	sleep 1
 
-	for l in `ls /dev/block/zram*`; do
+	for l in `ls /dev/block/zram*`
+	do
 		swapon $l
 	done
 
 	clear
-	echo "${yellow}zRAM enabled!${nc}"
+	echo "${yellow}zRAM enabled!$nc"
 	sleep 1
 	
 	zram_settings
@@ -563,15 +597,16 @@ zram_enable(){
 
 zram_disable(){
 	clear
-	echo "${yellow}Disabling zRAM...${nc}"
+	echo "${yellow}Disabling zRAM...$nc"
 	sleep 1
 
-	for l in `ls /dev/block/zram*`; do
+	for l in `ls /dev/block/zram*`
+	do
 		swapoff $l
 	done
 
 	clear
-	echo "${yellow}zRAM disabled!${nc}"
+	echo "${yellow}zRAM disabled!$nc"
 	sleep 1
 	
 	zram_settings
@@ -579,17 +614,19 @@ zram_disable(){
 
 catalyst_control(){
 	clear
-	echo "${yellow}Game Booster${nc}"
+	echo "${yellow}Game Booster$nc"
+	echo
 	echo "[1] Boost"
 	echo "[2] Options"
 	echo "[B] Back"
+	echo
 	echo -n "> "
 	read game_booster_opt
 	case $game_booster_opt in
 		1 ) catalyst_inject;;
 		2 ) catalyst_time_cfg;;
 		b|B ) body;;
-		* ) unknown_option && catalyst_control;;
+		* ) unknown_option; catalyst_control;;
 	esac
 }
 
@@ -597,9 +634,9 @@ catalyst_inject(){
 	clear
 	echo "Please leave the terminal emulator running"
 	echo "This will continue to run untill close the terminal"
-	echo
 
-	while true; do
+	while true
+	do
 		sync
   		echo "3" > /proc/sys/vm/drop_caches
 		sleep $catalyst_time
@@ -608,14 +645,14 @@ catalyst_inject(){
 
 catalyst_time_cfg(){
 	clear
-	echo "Current rate: $catalystsec"
+	echo "Current rate: $catalyst_time"
 	echo "60 - Every minute - Default"
 	echo "3600 - Every hour"
 	echo
 	echo "Please enter a rate in seconds:"
 	echo -n "> "
 	read catalyst_time_val
-	setprop hybrid.catalyst_time $catalyst_time_val;
+	setprop hybrid.catalyst_time $catalyst_time_val
 	clear
 	echo "Time updated!"
 	sleep 1
@@ -626,33 +663,34 @@ catalyst_time_cfg(){
 
 options(){
 	clear
-	echo "${yellow}How to install tweaks?${nc}"
+	echo "${yellow}How to install tweaks?$nc"
 	echo
 	echo " T|Temporary installs"
 	echo " P|Permanent installs"
-	first_run
-	echo -ne "> "
+	first_run_tip
+	echo
+	echo -n "> "
 	read usagetype_first_start_opt
 	case $usagetype_first_start_opt in
-		t|T ) setprop hybrid.perm 0 && echo "Ok!" && sleep 1;;
-		p|P ) setprop hybrid.perm 1 && echo "Ok!" && sleep 1;;
-		* ) unknown_option && options;;
+		t|T ) setprop hybrid.perm 0; clear; echo "Done"; sleep 1;;
+		p|P ) setprop hybrid.perm 1; clear; echo "Done"; sleep 1;;
+		* ) unknown_option; options;;
 	esac
 }
 
 about_info(){
 	clear
-	echo "${green}About:${nc}"
+	echo "${green}About:$nc"
 	echo
 	echo "Hybrid Version: $ver_revision"
 	echo
-	echo "${yellow}INFO${nc}"
+	echo "${yellow}INFO$nc"
 	echo "This script deals with many things apps normally do."
-	echo "But this script is ${cyan}AWESOME!${nc} because its < ${bld}1MB!${nc}"
+	echo "But this script is ${cyan}AWESOME!$nc because its < ${bld}1MB!$nc"
 	echo
-	echo "${yellow}CREDITS${nc}"
+	echo "${yellow}CREDITS$nc"
 	echo "DiamondBond : Script creator & maintainer"
-	echo "Hoholee12/Wedgess/Imbawind/Luca020400 : Code ${yellow}:)${nc}"
+	echo "Hoholee12/Wedgess/Imbawind/Luca020400 : Code $yellow:)$nc"
 	sleep 5
 }
 
@@ -669,7 +707,8 @@ custom_reboot(){
 	clear
 	echo "Bam!"
 	sleep 1
-	sync && reboot
+	sync
+	reboot
 }
 
 safe_exit(){
@@ -678,24 +717,28 @@ safe_exit(){
 	exit
 }
 
-first_run(){
+first_run_tip(){
 	if [ "$perm" = "" ]
 	then
  	 	echo
-	 	echo " You can change it in Options later"
+	 	echo " You can change it in Options later*"
 	fi
 }
 
 clear
-if [ $EUID -ne 0 ]; then #to be revised
+if [ $EUID -ne 0 ] #to be revised
+then
 	echo "This script must be run as root"
 	exit 1
-elif [ "$perm" = "" ]; then
+elif [ "$perm" = "" ]
+then
 	options
-elif [ "$catalyst_time" = "" ]; then
+elif [ "$catalyst_time" = "" ]
+then
 	setprop hybrid.catalyst_time 60
 fi
-while true; do
+while true
+do
 	mount -o remount rw /system >/dev/null 2>&1
 	body
 done
